requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via jsonschema
django==4.1.2
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-filter==22.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.2.1
    # via -r requirements.in
drf-spectacular==0.24.2
    # via -r requirements.in
drf-writable-nested==0.7.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.16.0
    # via drf-spectacular
pyjwt==2.5.0
    # via djangorestframework-simplejwt
pyrsistent==0.18.1
    # via jsonschema
pytz==2022.5
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via drf-spectacular
